requirements.txt
Flask==3.1.2
flask-cors==6.0.2
gunicorn==23.0.0
google-generativeai==0.8.6
python-dotenv==1.1.0
Werkzeug==3.1.3
psycopg2-binary==2.9.9
